backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
gunicorn>=21.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pymongo>=4.6.0
yfinance>=0.2.41
openai>=1.12.0
google-genai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
slowapi>=0.1.9
python-multipart>=0.0.9
email-validator>=2.1.0
resend>=0.8.0
pyotp>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
pandas>=2.2.0
numpy>=1.26.0
